backend/requirements.txt
flask
flask-cors
pymongo
qrcode
pillow